server/requirements.txt
flask
flask_login
flask_sqlalchemy
flask_json
sqlalchemy
pyjwt
werkzeug
requests
firebase-admin
elasticsearch
wtforms
wtforms_json